requirements.txt
streamlit
requests
python-dotenv
streamlit-geolocation
